converting all painting to double precision

+++ xchart/src/main/java/com/xeiam/xchart/internal/markers/Marker.java
@@ -30,5 +30,5 @@ public abstract class Marker {
   public static final double X_OFFSET = (int) (-1.0 * (SIZE / 2.0));
   public static final double Y_OFFSET = (int) (-1.0 * (SIZE / 2.0));
 
-  public abstract void paint(Graphics2D g, int xOffset, int yOffset);
+  public abstract void paint(Graphics2D g, double xOffset, double yOffset);
 }
